Brugeradmin opdateret

+++ AbonnementsimuleringKlient/SimuleringsVindue/SimuleringsVindue.cs
@@ -164,7 +164,7 @@ namespace AbonnementsimuleringKlient
             try
             {
                 label2.Hide();
-                var index = Convert.ToInt32(textBox1.Text);
+                var index = Convert.ToDecimal(textBox1.Text);
                 if (!this._arbejder.IsBusy)
                 {
                     KorNy.Enabled = false;
